Updated address autocomplete, Added base http resource, Updated resources
