+++ workspace/libs/shared/data-access/src/index.ts
@@ -1,5 +1,6 @@
 export * from './lib/shared-data-access.module';
 export * from './lib/interceptors';
+export * from './lib/interfaces';
 export * from './lib/models';
 export * from './lib/enums';
 export * from './lib/classes';
