+++ workspace/libs/shared/data-access/src/lib/classes/index.ts
@@ -1,2 +1,3 @@
 export * from './abstract-crud-resource';
+export * from './abstract-http-resource';
 export * from './abstract-resource';
