+++ workspace/libs/shared/data-access/src/lib/interfaces/crud-resource.interface.ts
@@ -1,6 +1,6 @@
 import { Observable } from 'rxjs';
 
-import { NoContent } from '../classes/abstract-resource';
+import { NoContent } from '../classes/abstract-http-resource';
 
 export interface ICrudResource<T> {
   create(id: number, payload: T): Observable<T | null>;
